feat: fixed f second review
