+++ src/components/burger-constructor/burger-constructor.module.css
@@ -12,7 +12,8 @@
     max-height: 464px;
 
     list-style-type: none;
-    overflow: scroll;
+    overflow-y: scroll;
+    overflow-x: hidden;
 }
 
 .item {
