+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -11,6 +11,7 @@
     margin: 0;
     padding: 0;
 
-    overflow: scroll;
+    overflow-y: scroll;
+    overflow-x: hidden;
     list-style-type: none;
 }
